First mostly working version
rrqueue-like queue object, organised relationship between contexts
and queues (though some confusion remains with the root).

Some disentangling to go as there is some general approaches here.  I suspect
that this is a special case of a "file system" based queuing system, with
some specialisations for dealing with the web API.